Refactored API query code
